In Contact, added explicit Enable() and Disable() methods. Typically, code will be passing in a constant true or false to the Enabled property, causing a needless branch in the propery code. Using the explicit methods eliminates the branch overhead.
In Contact.Update, fixed some subtle problems related to callback triggering when fixtures disabled the collision via return value from Fixture.OnCollision. Previously, a fixture that returned true from OnCollision would not have it's corresponding OnSeparation callback invoked if the other fixture disabled the contact within this loop. It would however get the callback later if the contact were disabled for other reasons. This change ensures that any fixture that returns true from OnCollision will have it's OnSeparation callback invoked when the contact is disabled for any reason within the scope of this function, making the behavior consistent with the other ways OnSeparation is invoked. Users can now reliably expect that OnSeparation will be invoked at some point when a contact becomes disabled, enabling scenarios where OnCollision sets up some data that can be expected to be cleaned up in OnSeparation. This problem only occurs when both fixtures have OnCollision callbacks that are able to return false, which is a fairly advanced use case I don't think people use very much or this would have been noticed sooner. Because it's a little complicated and makes the diff ugly, I left the original code intact but disabled via #if blocks for the time being. The real world use case that revealed this problem was a particle system that was activated when objects scraped into each other via OnCollision callback, but never turned off on separation because one of the objects had disabled the collision, denying the other object's OnSeparation callback the opportunity to detect the contact had become disabled.

git-svn-id: https://farseerphysics.svn.codeplex.com/svn@90041 81e6e737-5a20-4d0f-a97c-cb01297b46e0